Refactor named tensor helper code
A lot of these helper functions *should* be methods on TensorImpl, but
they're not due to working around some issues. All of these helper
functions should exist in NamedTensor.cpp. They'll be useful for
implementing named inference rules in TH (since tensors in TH are
TensorImpl*).

Test Plan
- [namedtensor ci]
