+++ aten/src/ATen/NamedTensor.h
@@ -7,10 +7,16 @@
 
 namespace at {
 
-// TensorImpl has a unique_ptr<NamedTensorMetaInterface> field. Ideally we would
-// just put optional<vector<Dimname>> into TensorImpl, but the problem with that is
-// that c10::Symbol isn't actually a part of the c10 lib (where TensorImpl is).
-// In the long term, we should decouple c10::Symbol from aten and toss it into c10.
+// XXX: This file exists because TensorImpl is in c10, but Dimname is in ATen.
+// Due to the c10/ATen library split, TensorImpl cannot depend on Dimname,
+// so we have a couple of workarounds.
+//
+// In the long term, we'll move Dimname to c10 and everything in this file
+// can be refactored out. The main blocker for that is that "c10::Symbol"
+// actually exists outside of c10 and needs to be moved in.
+
+// TensorImpl has a unique_ptr<NamedTensorMetaInterface> field.
+// XXX: Ideally we would just put optional<vector<Dimname>> into TensorImpl.
 struct CAFFE2_API NamedTensorMeta : public c10::NamedTensorMetaInterface {
   explicit NamedTensorMeta(int64_t num_names)
     : names_(std::vector<Dimname>(num_names, Dimname::wildcard())) {}
@@ -33,6 +39,17 @@ struct CAFFE2_API NamedTensorMeta : public c10::NamedTensorMetaInterface {
  private:
   std::vector<Dimname> names_;
 };
+
+namespace impl {
+
+// Some helper functions on TensorImpl. Useful for working with names in TH.
+// XXX: Ideally these would exist as methods on TensorImpl
+CAFFE2_API void internal_set_names_inplace(TensorImpl* impl, optional<DimnameList> names);
+CAFFE2_API optional<DimnameList> internal_get_names(TensorImpl* impl);
+CAFFE2_API bool internal_is_named(TensorImpl* impl);
+
+
+} // namespace impl
 
 } // namespace at
 #endif
