+++ aten/src/ATen/core/TensorMethods.h
@@ -1779,8 +1779,7 @@ inline const NamedTensorMeta* Tensor::get_named_tensor_meta() const {
 }
 
 inline bool Tensor::is_named() const {
-  auto* named_tensor_meta = get_named_tensor_meta();
-  return named_tensor_meta != nullptr && named_tensor_meta->has_names();
+  return impl::internal_is_named(unsafeGetTensorImpl());
 }
 #endif
